First cut of libfaasm working
