+++ lib/faasm/counter.h
@@ -1,7 +1,7 @@
 #ifndef FAASM_COUNTER_H
 #define FAASM_COUNTER_H
 
-#include "faasm/faasm.h"
+#include "faasm/memory.h"
 
 namespace faasm {
     void initCounter(FaasmMemory *memory, const char* counterKey);
