+++ lib/memory.cpp
@@ -1,5 +1,6 @@
-#include "faasm/faasm.h"
+#include "faasm/memory.h"
 
+extern "C" {
 long __faasm_read_state(const char *key, unsigned char *buffer, long bufferLen);
 
 void __faasm_write_state(const char *key, unsigned char *data, long dataLen);
@@ -13,10 +14,15 @@ long __faasm_read_input(unsigned char *buffer, long bufferLen);
 void __faasm_write_output(const unsigned char *output, long outputLen);
 
 void __faasm_chain_function(const char *name, const unsigned char *inputData, long inputDataSize);
+}
 
 namespace faasm {
 
-    FaasmMemory::FaasmMemory() = default;
+    static const int MAX_INPUT_BYTES = 65536;
+
+    FaasmMemory::FaasmMemory() {
+
+    }
 
     long FaasmMemory::getStateSize(const char *key) {
         uint8_t buf[1];
@@ -48,7 +54,7 @@ namespace faasm {
         return __faasm_read_input(buf, 0);
     }
 
-    const uint8_t* FaasmMemory::getInput() {
+    const uint8_t *FaasmMemory::getInput() {
         auto inputArray = new uint8_t[MAX_INPUT_BYTES];
         __faasm_read_input(inputArray, MAX_INPUT_BYTES);
 
@@ -67,4 +73,3 @@ namespace faasm {
         __faasm_chain_function(name, inputData, inputDataSize);
     }
 };
-
